feat: implement deferRoute utility for handling route decisions in tests
